+++ src/OpenRasta.DI.Windsor/ConvertLifestyles.cs
